Fixed SD2-565.  Cutting and pasting a file into its original directory no longer deletes the file.
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@771 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Main/Base/Project/Src/Gui/Pads/ProjectBrowser/TreeNodes/DirectoryNode.cs
@@ -458,7 +458,10 @@ namespace ICSharpCode.SharpDevelop.Project
 				if (File.Exists(clipboardObject.FileName)) {
 					ICSharpCode.SharpDevelop.Project.Commands.AddExistingItemsToProject.CopyFile(clipboardObject.FileName, this, true);
 					if (clipboardObject.PerformMove) {
-						FileService.RemoveFile(clipboardObject.FileName, false);
+						string copiedFileName = Path.Combine(Directory, Path.GetFileName(clipboardObject.FileName));
+						if (!FileUtility.IsEqualFileName(clipboardObject.FileName, copiedFileName)) {
+							FileService.RemoveFile(clipboardObject.FileName, false);
+						}
 					}
 				}
 			} else if (dataObject.GetDataPresent(typeof(DirectoryNode))) {
